+++ packages/orq-rc/RELEASES.md
@@ -11199,3 +11199,13 @@ Based on:
 - [python v4.16.0-rc.18] packages/orq-rc
 ### Releases
 - [PyPI v4.16.0-rc.18] https://pypi.org/project/orq-ai-sdk/4.16.0-rc.18 - packages/orq-rc
+
+## 2026-09-25 09:27:56
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.799.0 (2.941.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.16.0-rc.19] packages/orq-rc
+### Releases
+- [PyPI v4.16.0-rc.19] https://pypi.org/project/orq-ai-sdk/4.16.0-rc.19 - packages/orq-rc